Cleaning up the code and documenting important parts of the project using Ruff and Mypy

+++ source/backend/app/api/routes/auth.py
@@ -1,4 +1,4 @@
-from typing import Annotated
+from typing import Annotated  # noqa: I001
 
 from fastapi import APIRouter, Depends, Request, Response
 from fastapi.security import OAuth2PasswordRequestForm
@@ -7,13 +7,13 @@ from ..dependencies import UnitOfWorkContext
 from ... import schemas
 from ...core.settings import config
 from ...services import AuthenticationService
-from ...utils.exceptions import InvalidCredentialsException
+from ...utils.exceptions import InvalidTokenException
 
 router = APIRouter()
 
 
 @router.post("/login", response_model=schemas.Token)
-async def login(
+async def access_token(
     response: Response,
     uow: UnitOfWorkContext,
     *,
@@ -22,8 +22,6 @@ async def login(
     user = await AuthenticationService.authenticate_user(
         uow, email=credentials.username, password=credentials.password
     )
-    if not user:
-        raise InvalidCredentialsException
 
     token = await AuthenticationService.create_token(uow, user.id)
 
@@ -37,28 +35,21 @@ async def login(
     return schemas.Token(
         access_token=token.access_token,
         refresh_token=token.refresh_token,
-        token_type="bearer",
+        token_type=config.Authentication().TOKEN_TYPE,
     )
 
 
 @router.post("/refresh", response_model=schemas.Token)
-async def login(
+async def refresh_access_token(
     request: Request,
-    response: Response,
     uow: UnitOfWorkContext,
 ) -> schemas.Token:
-    refreshed_token = await AuthenticationService.refresh_token(
-        uow, refresh_token=request.cookies.get("refresh_token")
-    )
+    refresh_token = request.cookies.get("refresh_token")
 
-    response.set_cookie(
-        "refresh_token",
-        refreshed_token.refresh_token,
-        max_age=config.Authentication().REFRESH_TOKEN_EXPIRE_DAYS * 60 * 60 * 24,
-        httponly=True,
-    )
+    if not refresh_token:
+        raise InvalidTokenException
 
-    return refreshed_token
+    return await AuthenticationService.refresh_token(uow, refresh_token=refresh_token)
 
 
 @router.post("/logout")
@@ -69,8 +60,9 @@ async def logout(
 ) -> None:
     refresh_token = request.cookies.get("refresh_token")
 
+    if not refresh_token:
+        raise InvalidTokenException
+
     response.delete_cookie("refresh_token", httponly=True)
 
     await AuthenticationService.logout(uow, refresh_token=refresh_token)
-
-    return None
